rename AdaptiveViolation to AdaptiveWarning
