+++ source/dotnet/Library/AdaptiveCards/AdaptiveException.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace AdaptiveCards
+{
+    public class AdaptiveException : Exception
+    {
+        public AdaptiveException()
+        {
+        }
+
+        public AdaptiveException(string message) : base(message)
+        {
+        }
+
+        public AdaptiveException(string message, Exception innerException) : base(message, innerException)
+        {
+        }
+    }
+}
