+++ source/dotnet/Library/AdaptiveCards/AdaptiveSerializationException.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace AdaptiveCards
+{
+    public class AdaptiveSerializationException : AdaptiveException
+    {
+        public AdaptiveSerializationException()
+        {
+        }
+
+        public AdaptiveSerializationException(string message) : base(message)
+        {
+        }
+
+        public AdaptiveSerializationException(string message, Exception innerException) : base(message, innerException)
+        {
+        }
+    }
+}
